Simplified UObject structs' hierarchy
