+++ src/main/kotlin/me/fungames/jfortniteparse/fileprovider/FileProvider.kt
@@ -4,8 +4,8 @@ import me.fungames.jfortniteparse.exceptions.ParserException
 import me.fungames.jfortniteparse.ue4.assets.Package
 import me.fungames.jfortniteparse.ue4.locres.FnLanguage
 import me.fungames.jfortniteparse.ue4.locres.Locres
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FPackageIndex
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FSoftObjectPath
+import me.fungames.jfortniteparse.ue4.objects.uobject.FPackageIndex
+import me.fungames.jfortniteparse.ue4.objects.uobject.FSoftObjectPath
 import me.fungames.jfortniteparse.ue4.pak.GameFile
 import me.fungames.jfortniteparse.ue4.registry.AssetRegistry
 import me.fungames.jfortniteparse.ue4.versions.Ue4Version
