+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/exports/UStringTable.kt
@@ -2,8 +2,8 @@ package me.fungames.jfortniteparse.ue4.assets.exports
 
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FObjectExport
+import me.fungames.jfortniteparse.ue4.objects.uobject.FName
+import me.fungames.jfortniteparse.ue4.objects.uobject.FObjectExport
 
 
 typealias FMetaDataMap = MutableMap<FName, String>
