+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/exports/mats/UMaterial.kt
@@ -4,7 +4,7 @@ import me.fungames.jfortniteparse.converters.ue4.CMaterialParams
 import me.fungames.jfortniteparse.ue4.assets.enums.EBlendMode
 import me.fungames.jfortniteparse.ue4.assets.exports.tex.UTexture
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FObjectExport
+import me.fungames.jfortniteparse.ue4.objects.uobject.FObjectExport
 import me.fungames.jfortniteparse.ue4.versions.GAME_UE4_BASE
 
 @ExperimentalUnsignedTypes
