+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/exports/mats/UMaterialInstance.kt
@@ -2,7 +2,7 @@ package me.fungames.jfortniteparse.ue4.assets.exports.mats
 
 import me.fungames.jfortniteparse.ue4.assets.objects.structs.FMaterialInstanceBasePropertyOverrides
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FObjectExport
+import me.fungames.jfortniteparse.ue4.objects.uobject.FObjectExport
 
 @ExperimentalUnsignedTypes
 open class UMaterialInstance(Ar: FAssetArchive, exportObject: FObjectExport) : UMaterialInterface(Ar, exportObject) {
