+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/exports/tex/UTexture.kt
@@ -2,7 +2,7 @@ package me.fungames.jfortniteparse.ue4.assets.exports.tex
 
 import me.fungames.jfortniteparse.converters.ue4.CMaterialParams
 import me.fungames.jfortniteparse.ue4.assets.exports.mats.UUnrealMaterial
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FObjectExport
+import me.fungames.jfortniteparse.ue4.objects.uobject.FObjectExport
 
 @ExperimentalUnsignedTypes
 abstract class UTexture(export: FObjectExport) : UUnrealMaterial(export) {
