+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/exports/valorant/CharacterRoleDataAsset.kt
@@ -5,8 +5,8 @@ import me.fungames.jfortniteparse.ue4.assets.exports.UObject
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
 import me.fungames.jfortniteparse.ue4.objects.core.misc.FGuid
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FObjectExport
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FSoftObjectPath
+import me.fungames.jfortniteparse.ue4.objects.uobject.FObjectExport
+import me.fungames.jfortniteparse.ue4.objects.uobject.FSoftObjectPath
 
 @ExperimentalUnsignedTypes
 class CharacterRoleDataAsset : UExport {
