+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/objects/FPropertyTag.kt
@@ -5,7 +5,7 @@ import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
 import me.fungames.jfortniteparse.ue4.objects.core.misc.FGuid
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
+import me.fungames.jfortniteparse.ue4.objects.uobject.FName
 
 @ExperimentalUnsignedTypes
 class FPropertyTag : UClass {
