+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/objects/FStructFallback.kt
@@ -4,7 +4,7 @@ import me.fungames.jfortniteparse.exceptions.ParserException
 import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
+import me.fungames.jfortniteparse.ue4.objects.uobject.FName
 
 @ExperimentalUnsignedTypes
 class FStructFallback : UClass {
