+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/objects/structs/FScalarParameterValue.kt
@@ -2,7 +2,7 @@ package me.fungames.jfortniteparse.ue4.assets.objects.structs
 
 import me.fungames.jfortniteparse.ue4.assets.util.StructFallbackClass
 import me.fungames.jfortniteparse.ue4.assets.util.StructFieldName
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
+import me.fungames.jfortniteparse.ue4.objects.uobject.FName
 
 @StructFallbackClass
 class FScalarParameterValue(
