+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/objects/gameplaytags/GameplayTagContainer.kt
@@ -3,7 +3,7 @@ package me.fungames.jfortniteparse.ue4.objects.gameplaytags
 import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
+import me.fungames.jfortniteparse.ue4.objects.uobject.FName
 
 @ExperimentalUnsignedTypes
 class FGameplayTagContainer : UClass {
