+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/objects/uobject/ObjectResource.kt
@@ -1,4 +1,4 @@
-package me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject
+package me.fungames.jfortniteparse.ue4.objects.uobject
 
 import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.assets.Package
@@ -6,6 +6,7 @@ import me.fungames.jfortniteparse.ue4.assets.exports.UExport
 import me.fungames.jfortniteparse.ue4.assets.reader.FAssetArchive
 import me.fungames.jfortniteparse.ue4.assets.writer.FAssetArchiveWriter
 import me.fungames.jfortniteparse.ue4.objects.core.misc.FGuid
+import me.fungames.jfortniteparse.ue4.objects.uobject.FName
 import me.fungames.jfortniteparse.ue4.versions.*
 import me.fungames.jfortniteparse.ue4.writer.FArchiveWriter
