+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/objects/uobject/UInterfaceProperty.kt
@@ -1,4 +1,4 @@
-package me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject
+package me.fungames.jfortniteparse.ue4.objects.uobject
 
 import me.fungames.jfortniteparse.ue4.UClass
 import me.fungames.jfortniteparse.ue4.reader.FArchive
