+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/registry/objects/FAssetIdentifier.kt
@@ -1,7 +1,7 @@
 package me.fungames.jfortniteparse.ue4.registry.objects
 
 import me.fungames.jfortniteparse.ue4.UClass
-import me.fungames.jfortniteparse.ue4.objects.coreuobject.uobject.FName
+import me.fungames.jfortniteparse.ue4.objects.uobject.FName
 import me.fungames.jfortniteparse.ue4.registry.reader.FNameTableArchive
 
 @ExperimentalUnsignedTypes
